.
Zavrnil dostop navadnemu uporabniku.
Stran je dostopna le adminu

+++ Includes/delete.inc.php
@@ -4,9 +4,19 @@ require 'dbh.inc.php';
 
 $idNarocilo = $_GET['idTaco'];
 
-$sql = "DELETE FROM narocila WHERE idNarocilo = $idNarocilo";
-$result = mysqli_query($conn, $sql);
+if (isset($idNarocilo)) {
+  $sql = "DELETE FROM narocila WHERE idNarocilo = $idNarocilo";
+  $result = mysqli_query($conn, $sql);
+  header("Location: ../pregledNarocil.php?order=delete");
+}
 
-header("Location: ../pregledNarocil.php?order=delete");
+$idKosarica = $_GET['idKosarica'];
+
+if (isset($idKosarica)) {
+  $sql = "DELETE FROM kosarica WHERE idKosarica = $idKosarica";
+  $result = mysqli_query($conn, $sql);
+  header("Location: ../pregledNarocil.php?order=delete");
+}
+
 
 ?>
